Create classification_models.ipynb
Full notebook script for my capstone project at CUNY SPS in detecting school shooters according to their digital writings. This entails looking for school shooters by using Doc2Vec's PV-BDOW and DM (paragraph vector distributed bag of words and distributed memory) as features, then using those features within classifiers. Logistic Regression and MLPNNet performed the best with the combination of these features under a ratio of non-shooters to shooters, 96% to 4% respectively. However, it is import to note the classifiers' performances do not withstand in identifying school shooters under a higher ratio of non-shooters to shooters (99.66% to 0.44%), as the identification of shooters drastically dwindles according to the steps taken in this script.